msg template refactored
