+++ src/matrixbot.ts
@@ -1,28 +1,25 @@
 import got from 'got';
 
-import { TransactionData, Notifier } from './types';
+import { TransactionData, Notifier, MatrixbotMsg } from './types';
 
-interface LabelMap {
-    alertname: string;
-    severity: string;
-}
+export const MsgTemplate = {
+    "receiver": "webhook",
+    "status": "firing",
+    "alerts": [
+        {
+            "status": "firing",
+            "labels": {
+                "alertname": "TransactionSent",
+                "severity": "info"
+            },
+            "annotations": {
+                "description": ""
+            }
+        }
+    ],
+    "version": "4"
+};
 
-interface Annotation {
-    description: string;
-}
-
-interface Alert {
-    status: string;
-    labels: LabelMap;
-    annotations: Annotation;
-}
-
-interface MatrixbotMsg {
-    receiver: string;
-    status: string;
-    alerts: Array<Alert>;
-    version: string;
-}
 
 export class Matrixbot implements Notifier {
     constructor(private readonly endpoint: string) { }
@@ -34,23 +31,11 @@ export class Matrixbot implements Notifier {
     }
 
     private transactionMsg(data: TransactionData): MatrixbotMsg {
-        return {
-            "receiver": "webhook",
-            "status": "firing",
-            "alerts": [
-                {
-                    "status": "firing",
-                    "labels": {
-                        "alertname": "TransactionSent",
-                        "severity": "info"
-                    },
-                    "annotations": {
-                        "description": `New transaction sent from account ${data.name}, check https://polkascan.io/pre/${data.networkId}/account/${data.address}#transactions for details`
-                    }
-                }
-            ],
-            "version": "4"
-        };
+        const msg = { ...MsgTemplate };
+
+        msg.alerts[0].annotations.description = `New transaction sent from account ${data.name}, check https://polkascan.io/pre/${data.networkId}/account/${data.address}#transactions for details`;
+
+        return msg;
     }
 
     private async send(json: MatrixbotMsg): Promise<string> {
