+++ src/types.ts
@@ -29,10 +29,38 @@ export interface PromClient {
     resetTotalValidatorOfflineReports(name: string): void;
 }
 
+export enum TransactionType {
+    Received,
+    Sent
+}
+
 export interface TransactionData extends Subscribable {
+    txType: TransactionType;
     networkId: string;
 }
 
 export interface Notifier {
     newTransaction(data: TransactionData): Promise<string>;
+}
+
+interface LabelMap {
+    alertname: string;
+    severity: string;
+}
+
+interface Annotation {
+    description: string;
+}
+
+interface Alert {
+    status: string;
+    labels: LabelMap;
+    annotations: Annotation;
+}
+
+export interface MatrixbotMsg {
+    receiver: string;
+    status: string;
+    alerts: Array<Alert>;
+    version: string;
 }
